requirements.txt
ultralytics>=8.3.0
pyyaml>=6.0.1
torch>=2.11.0
torchvision>=0.26.0
torchaudio>=2.11.0
pillow-heif>=1.1.1
psutil>=7.1.3
